+++ Nickel/Framework/Implementations/Content/ModCards.cs
@@ -1,6 +1,7 @@
 using Nickel.Models.Content;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Nickel;
 
@@ -135,6 +136,118 @@ internal sealed class ModCards(
 	{
 		add => cardTraitManagerProvider().OnSetCardTraitOverrideEvent.Add(value, modManifest);
 		remove => cardTraitManagerProvider().OnSetCardTraitOverrideEvent.Remove(value, modManifest);
+	}
+
+	public bool IsCurrentlyCreatingCardTraitStates(State state, Card card)
+		=> cardTraitManagerProvider().IsCurrentlyCreatingCardTraitStates(card);
+}
+
+internal sealed class VanillaModCards(
+	Func<CardManager> cardManagerProvider,
+	Func<CardTraitManager> cardTraitManagerProvider
+) : IModCards
+{
+	private readonly Lazy<Dictionary<string, ICardEntry>> LazyRegisteredCards = new(() => DB.cards.Where(kvp => kvp.Value.Assembly == typeof(Card).Assembly).ToDictionary(kvp => kvp.Key, kvp => cardManagerProvider().LookupByCardType(kvp.Value)!));
+	private readonly Lazy<Dictionary<string, ICardTraitEntry>> LazyRegisteredCardTraits = new(() =>
+	{
+		List<ICardTraitEntry> traits = [
+			cardTraitManagerProvider().ExhaustCardTrait,
+			cardTraitManagerProvider().RetainCardTrait,
+			cardTraitManagerProvider().RecycleCardTrait,
+			cardTraitManagerProvider().InfiniteCardTrait,
+			cardTraitManagerProvider().UnplayableCardTrait,
+			cardTraitManagerProvider().TemporaryCardTrait,
+			cardTraitManagerProvider().BuoyantCardTrait,
+			cardTraitManagerProvider().SingleUseCardTrait,
+		];
+		return traits.ToDictionary(e => e.UniqueName);
+	});
+	
+	public IReadOnlyDictionary<string, ICardEntry> RegisteredCards
+		=> this.LazyRegisteredCards.Value;
+	
+	public IReadOnlyDictionary<string, ICardTraitEntry> RegisteredTraits
+		=> this.LazyRegisteredCardTraits.Value;
+	
+	public ICardEntry? LookupByCardType(Type cardType)
+		=> cardManagerProvider().LookupByCardType(cardType);
+
+	public ICardEntry? LookupByUniqueName(string uniqueName)
+		=> cardManagerProvider().LookupByUniqueName(uniqueName);
+
+	public ICardEntry RegisterCard(CardConfiguration configuration)
+		=> throw new NotSupportedException();
+
+	public ICardEntry RegisterCard(string name, CardConfiguration configuration)
+		=> throw new NotSupportedException();
+
+	public ICardTraitEntry ExhaustCardTrait
+		=> cardTraitManagerProvider().ExhaustCardTrait;
+
+	public ICardTraitEntry RetainCardTrait
+		=> cardTraitManagerProvider().RetainCardTrait;
+
+	public ICardTraitEntry RecycleCardTrait
+		=> cardTraitManagerProvider().RecycleCardTrait;
+
+	public ICardTraitEntry InfiniteCardTrait
+		=> cardTraitManagerProvider().InfiniteCardTrait;
+
+	public ICardTraitEntry UnplayableCardTrait
+		=> cardTraitManagerProvider().UnplayableCardTrait;
+
+	public ICardTraitEntry TemporaryCardTrait
+		=> cardTraitManagerProvider().TemporaryCardTrait;
+
+	public ICardTraitEntry BuoyantCardTrait
+		=> cardTraitManagerProvider().BuoyantCardTrait;
+
+	public ICardTraitEntry SingleUseCardTrait
+		=> cardTraitManagerProvider().SingleUseCardTrait;
+	
+	public ICardTraitEntry? LookupTraitByUniqueName(string uniqueName)
+		=> cardTraitManagerProvider().LookupByUniqueName(uniqueName);
+
+	public ICardTraitEntry RegisterTrait(string name, CardTraitConfiguration configuration)
+		=> throw new NotSupportedException();
+
+	public IReadOnlySet<ICardTraitEntry> GetActiveCardTraits(State state, Card card)
+		=> cardTraitManagerProvider().GetActiveCardTraits(state, card);
+
+	public IReadOnlyDictionary<ICardTraitEntry, CardTraitState> GetAllCardTraits(State state, Card card)
+		=> cardTraitManagerProvider().GetAllCardTraits(state, card);
+
+	public bool IsCardTraitActive(State state, Card card, ICardTraitEntry trait)
+		=> cardTraitManagerProvider().IsCardTraitActive(state, card, trait);
+
+	public CardTraitState GetCardTraitState(State state, Card card, ICardTraitEntry trait)
+		=> cardTraitManagerProvider().GetCardTraitState(state, card, trait);
+
+	public void SetCardTraitOverride(State state, Card card, ICardTraitEntry trait, bool? overrideValue, bool permanent)
+		=> throw new NotSupportedException();
+
+	public event EventHandler<GetVolatileCardTraitOverridesEventArgs>? OnGetVolatileCardTraitOverrides
+	{
+		add => throw new NotSupportedException();
+		remove => throw new NotSupportedException();
+	}
+	
+	public event EventHandler<GetDynamicInnateCardTraitOverridesEventArgs>? OnGetDynamicInnateCardTraitOverrides
+	{
+		add => throw new NotSupportedException();
+		remove => throw new NotSupportedException();
+	}
+	
+	public event EventHandler<GetFinalDynamicCardTraitOverridesEventArgs>? OnGetFinalDynamicCardTraitOverrides
+	{
+		add => throw new NotSupportedException();
+		remove => throw new NotSupportedException();
+	}
+	
+	public event EventHandler<SetCardTraitOverrideEventArgs>? OnSetCardTraitOverride
+	{
+		add => throw new NotSupportedException();
+		remove => throw new NotSupportedException();
 	}
 
 	public bool IsCurrentlyCreatingCardTraitStates(State state, Card card)