+++ Nickel/Framework/Implementations/Content/ModEnemies.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Nickel;
 
@@ -32,4 +33,31 @@ internal sealed class ModEnemies(
 		this.RegisteredEnemyStorage[name] = entry;
 		return entry;
 	}
+}
+
+internal sealed class VanillaModEnemies(
+	Func<EnemyManager> enemyManagerProvider
+) : IModEnemies
+{
+	private readonly Lazy<Dictionary<string, IEnemyEntry>> LazyRegisteredEnemies = new(
+		() => typeof(AI).Assembly.GetTypes()
+			.Where(t => t.IsAssignableTo(typeof(AI)) && t != typeof(AI))
+			.Select(t => enemyManagerProvider().LookupByEnemyType(t)!)
+			.ToDictionary(e => e.UniqueName)
+	);
+	
+	public IReadOnlyDictionary<string, IEnemyEntry> RegisteredEnemies
+		=> this.LazyRegisteredEnemies.Value;
+
+	public IEnemyEntry? LookupByEnemyType(Type enemyType)
+		=> enemyManagerProvider().LookupByEnemyType(enemyType);
+
+	public IEnemyEntry? LookupByUniqueName(string uniqueName)
+		=> enemyManagerProvider().LookupByUniqueName(uniqueName);
+
+	public IEnemyEntry RegisterEnemy(EnemyConfiguration configuration)
+		=> throw new NotSupportedException();
+
+	public IEnemyEntry RegisterEnemy(string name, EnemyConfiguration configuration)
+		=> throw new NotSupportedException();
 }