+++ Nickel/Framework/Implementations/ModUtilities.cs
@@ -36,3 +36,35 @@ internal sealed class ModUtilities(
 	public void ApplyDelayedHarmonyPatches()
 		=> delayedHarmonyManager.ApplyDelayedPatches();
 }
+
+internal sealed class VanillaModUtilities(
+	EnumCasePool enumCasePool,
+	IProxyManager<string> proxyManager,
+	DelayedHarmonyManager delayedHarmonyManager
+) : IModUtilities
+{
+	public T ObtainEnumCase<T>() where T : struct, Enum
+		=> enumCasePool.ObtainEnumCase<T>();
+
+	public void FreeEnumCase<T>(T @case) where T : struct, Enum
+		=> enumCasePool.FreeEnumCase(@case);
+
+	public IProxyManager<string> ProxyManager { get; } = proxyManager;
+
+	public object Unproxy(object potentialProxy)
+	{
+		var current = potentialProxy;
+		while (current is IProxyObject.IWithProxyTargetInstanceProperty proxyObject)
+			current = proxyObject.ProxyTargetInstance;
+		return current;
+	}
+
+	public IHarmony Harmony
+		=> throw new NotSupportedException();
+	
+	public IHarmony DelayedHarmony
+		=> throw new NotSupportedException();
+	
+	public void ApplyDelayedHarmonyPatches()
+		=> delayedHarmonyManager.ApplyDelayedPatches();
+}